// packages/sui/src/graphql/vendor/eventsource-parser/types.ts
export interface EventSourceMessage {
	/**
	 * The event type sent from the server. Note that this differs from the browser `EventSource`
	 * implementation in that browsers will default this to `message`, whereas this parser will
	 * leave this as `undefined` if not explicitly declared.
	 */
	event?: string | undefined;

	/**
	 * ID of the message, if any was provided by the server. Can be used by clients to keep the
	 * last received message ID in sync when reconnecting.
	 */
	id?: string | undefined;

	/**
	 * The data received for this message
	 */
	data: string;
}